Move db stuff under trader. Fix formatting etc. Add a few placeholder things for strategies and data sources.
