+++ apps/trader/backtest.ts
@@ -1,0 +1,6 @@
+import { makeAlpacaClient } from './factory/alpacaClientFactory.ts';
+
+if (import.meta.main) {
+    const client = makeAlpacaClient();
+    console.log(client);
+}
