+++ apps/trader/dataProvider/dailyOhlcvData.ts
@@ -1,0 +1,7 @@
+import { DataProvider } from './dataProvider.ts';
+
+export class DailyOhlcvDataProvider extends DataProvider {
+    start() {
+        console.log('Starting daily OHLCV data provider');
+    }
+}
