+++ apps/trader/dataProvider/dataProvider.ts
@@ -1,0 +1,3 @@
+export abstract class DataProvider {
+    abstract start(): Promise<void> | void;
+}
