+++ apps/trader/deno.json
@@ -2,9 +2,14 @@
     "name": "@trader-seal/trader",
     "exports": "./main.ts",
     "tasks": {
-        "dev": "deno run --inspect --check --allow-net=localhost,paper-api.alpaca.markets,data.alpaca.markets --allow-env --env-file main.ts"
+        "dev": "deno run --check --allow-net=localhost,paper-api.alpaca.markets,data.alpaca.markets --allow-env --env-file main.ts",
+        "backtest": "deno run --check --allow-net=localhost,paper-api.alpaca.markets,data.alpaca.markets --allow-env --env-file backtest.ts",
+        "migrate-latest": "deno run --allow-read=migrations --allow-net=127.0.0.1:5432 --env-file --allow-env migrate.ts"
     },
     "imports": {
-        "@std/assert": "jsr:@std/assert@1"
+        "@std/assert": "jsr:@std/assert@1",
+        "kysely": "npm:kysely@^0.27.4",
+        "kysely-postgres-js": "npm:kysely-postgres-js@^2.0.0",
+        "postgres": "https://deno.land/x/postgresjs/mod.js"
     }
 }
