--- apps/yahoo-history-builder/dataToDb.ts
@@ -1,5 +1,0 @@
-import { db } from './db.ts';
-
-const assetsFile = await Deno.readTextFile('./input/assets.json');
-const assets = JSON.parse(assetsFile);
-
