+++ apps/yahoo-history-builder/deno.json
@@ -2,18 +2,9 @@
     "name": "@trader-seal/yahoo-history-builder",
     "exports": "./main.ts",
     "tasks": {
-        "scrape": "deno run --check --allow-read=output,input --allow-write=output --allow-net=query1.finance.yahoo.com scrapeYahoo.ts",
-        "save-db": "deno run --check --allow-read=output --allow-net=localhost:5432 --allow-env saveToDb.ts",
-        "migrate-latest": "deno run --allow-read=migrations --allow-net=127.0.0.1:5432 --env-file --allow-env migrate.ts"
+        "scrape": "deno run --check --allow-read=output,input --allow-write=output --allow-net=query1.finance.yahoo.com scrapeYahoo.ts"
     },
     "fmt": {
         "exclude": ["input", "output"]
-    },
-    "imports": {
-        "kysely": "npm:kysely@^0.27.4",
-        "kysely-postgres-js": "npm:kysely-postgres-js@^2.0.0",
-        "postgres": "https://deno.land/x/postgresjs/mod.js",
-        "pg": "npm:pg@^8.13.1",
-        "pg-pool": "npm:pg-pool@^3.7.0"
     }
 }
